KULRICE-4771 - Show/Hide buttons affect incorrect rows on Inquiry with multiple collections
The reason for this was that was that multiple subtabs had the same subtab id, so the show hide button would show or hide the first instance of the div id on the page.

I updated rowDisplay to use the current tab index to be the div ids so there aren't duplicate ids.

git-svn-id: https://svn.kuali.org/repos/rice/branches/rice-release-1-0-3-1-br@15179 7a7aa7f6-c479-11dc-97e2-85a2497f191d